refactoring NDocUtil
