+++ src/NDocUtil/DocUtil.cs
@@ -8,6 +8,10 @@ using System.Runtime.CompilerServices;
 using System.Text.RegularExpressions;
 using ColorCode;
 using JetBrains.Annotations;
+using NDocUtil.intern;
+using NDocUtil.intern.Console;
+using NDocUtil.intern.Snippets;
+using NDocUtil.intern.Winforms;
 using TheArtOfDev.HtmlRenderer.WinForms;
 
 namespace NDocUtil
@@ -91,25 +95,7 @@ namespace NDocUtil
                     string divSnippet = mCodeColorizer.Colorize(snippet.Body, language);
                     string htmlSnippet = string.Format(SNIPPET_HTML_TEMPLATE, divSnippet);
                     string snippetFilePath = BuildSnippetFilePath(callerFilePath, snippet.Name, "." + imageFormat.ToString().ToLower());
-
-                    Metafile image;
-
-                    IntPtr dib;
-                    var memoryHdc = Win32Utils.CreateMemoryHdc(IntPtr.Zero, 1, 1, out dib);
-                    try
-                    {
-                        image = new Metafile(memoryHdc, EmfType.EmfPlusDual, "..");
-
-                        using (var g = Graphics.FromImage(image))
-                        {
-                            HtmlRender.Render(g, htmlSnippet);
-                        }
-                    }
-                    finally
-                    {
-                        Win32Utils.ReleaseMemoryHdc(memoryHdc, dib);
-                    }
-
+                    Metafile image = HtmlToWmfUtil.Convert(htmlSnippet);
                     image.SaveAsEmf(snippetFilePath);
 
                 });
