+++ src/NDocUtil/NDocUtil.csproj
@@ -59,15 +59,20 @@
     <Compile Include="..\Version\src\NGitVersion\Generated\GlobalAssemblyInfo.cs">
       <Link>Properties\GlobalAssemblyInfo.cs</Link>
     </Compile>
-    <Compile Include="ConsoleMirroring.cs" />
-    <Compile Include="GraphicsUtil.cs" />
+    <Compile Include="intern\WinApi\BitMapInfo.cs" />
+    <Compile Include="intern\WinApi\BlendFunction.cs" />
+    <Compile Include="intern\Console\ConsoleMirroring.cs" />
+    <Compile Include="intern\Winforms\HtmlToWmfUtil.cs" />
+    <Compile Include="intern\Winforms\MetafileExtensions.cs" />
+    <Compile Include="intern\WinApi\TextMetric.cs" />
+    <Compile Include="intern\WinApi\Win32Utils.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="ConsoleRecord.cs" />
-    <Compile Include="ConsoleRecorder.cs" />
-    <Compile Include="Snippet.cs" />
+    <Compile Include="intern\Console\ConsoleRecord.cs" />
+    <Compile Include="intern\Console\ConsoleRecorder.cs" />
+    <Compile Include="intern\Snippets\Snippet.cs" />
     <Compile Include="DocUtil.cs" />
-    <Compile Include="SnippetParser.cs" />
-    <Compile Include="SnippetType.cs" />
+    <Compile Include="intern\Snippets\SnippetParser.cs" />
+    <Compile Include="intern\Snippets\SnippetType.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\NUtil\NUtil.csproj">
