+++ src/NDocUtil/intern/Console/ConsoleMirroring.cs
@@ -3,7 +3,7 @@ using System.IO;
 using System.Text;
 using JetBrains.Annotations;
 
-namespace NDocUtil
+namespace NDocUtil.intern.Console
 {
     /// <summary>inspired by: http://stackoverflow.com/questions/420429/mirroring-console-output-to-a-file </summary>
     public class ConsoleMirroring : TextWriter
@@ -15,23 +15,23 @@ namespace NDocUtil
 
         public ConsoleMirroring()
         {
-            if (Console.Out == null) throw new ArgumentException("Console.Out is null");
-            if (Console.Error == null) throw new ArgumentException("Console.Error is null");
+            if (System.Console.Out == null) throw new ArgumentException("Console.Out is null");
+            if (System.Console.Error == null) throw new ArgumentException("Console.Error is null");
 
             mTextWriter = new StringWriter();
-            mConsoleOutput = Console.Out;
-            mConsoleError = Console.Error;
+            mConsoleOutput = System.Console.Out;
+            mConsoleError = System.Console.Error;
 
-            Console.SetOut(this);
-            Console.SetError(this);
+            System.Console.SetOut(this);
+            System.Console.SetError(this);
         }
 
         protected override void Dispose(bool disposing)
         {
             if (disposing)
             {
-                Console.SetOut(mConsoleOutput);
-                Console.SetError(mConsoleError);
+                System.Console.SetOut(mConsoleOutput);
+                System.Console.SetError(mConsoleError);
             }
         }
