+++ src/NDocUtil/intern/WinApi/BitMapInfo.cs
@@ -1,0 +1,24 @@
+using System.Runtime.InteropServices;
+
+namespace NDocUtil.intern.WinApi
+{
+    [StructLayout(LayoutKind.Sequential)]
+    internal struct BitMapInfo
+    {
+        public int biSize;
+        public int biWidth;
+        public int biHeight;
+        public short biPlanes;
+        public short biBitCount;
+        public int biCompression;
+        public int biSizeImage;
+        public int biXPelsPerMeter;
+        public int biYPelsPerMeter;
+        public int biClrUsed;
+        public int biClrImportant;
+        public byte bmiColors_rgbBlue;
+        public byte bmiColors_rgbGreen;
+        public byte bmiColors_rgbRed;
+        public byte bmiColors_rgbReserved;
+    }
+}
