+++ src/NDocUtil/intern/WinApi/BlendFunction.cs
@@ -1,0 +1,21 @@
+using System.Runtime.InteropServices;
+
+namespace NDocUtil.intern.WinApi
+{
+    [StructLayout(LayoutKind.Sequential)]
+    internal struct BlendFunction
+    {
+        public byte BlendOp;
+        public byte BlendFlags;
+        public byte SourceConstantAlpha;
+        public byte AlphaFormat;
+
+        public BlendFunction(byte alpha)
+        {
+            BlendOp = 0;
+            BlendFlags = 0;
+            AlphaFormat = 0;
+            SourceConstantAlpha = alpha;
+        }
+    }
+}
